+++ r-naming.tf
@@ -1,0 +1,19 @@
+resource "azurecaf_name" "lb" {
+  name          = var.stack
+  resource_type = "azurerm_lb"
+  prefixes      = var.name_prefix == "" ? null : [local.name_prefix]
+  suffixes      = compact([var.client_name, var.location_short, var.environment, local.name_suffix, var.use_caf_naming ? "" : "lb"])
+  use_slug      = var.use_caf_naming
+  clean_input   = true
+  separator     = "-"
+}
+
+resource "azurecaf_name" "pubip" {
+  name          = var.stack
+  resource_type = "azurerm_public_ip"
+  prefixes      = var.name_prefix == "" ? null : [local.name_prefix]
+  suffixes      = compact([var.client_name, var.location_short, var.environment, local.name_suffix, var.use_caf_naming ? "" : "pubip"])
+  use_slug      = var.use_caf_naming
+  clean_input   = true
+  separator     = "-"
+}