+++ variables.tf
@@ -52,24 +52,6 @@ variable "ip_extra_tags" {
   default     = {}
 }
 
-variable "name_prefix" {
-  description = "Optional prefix for the generated name"
-  type        = string
-  default     = ""
-}
-
-variable "lb_custom_name" {
-  description = "Name of the Load Balancer, generated if not set."
-  type        = string
-  default     = ""
-}
-
-variable "ip_custom_name" {
-  description = "Name of the Public IP address, generated if not set."
-  type        = string
-  default     = ""
-}
-
 variable "allocate_public_ip" {
   description = "True to allocate a Public IP to the Load Balancer."
   type        = bool