+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Unreleased
+
+Breaking
+  * AZ-515: Option to use Azure CAF naming provider to name resources
+  * AZ-515: Require Terraform 0.13+
+
 # v4.1.0 - 2021-11-23
 
 Changed